refactor using ViewModel
