+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/features/accountonboarding/AccountOnboardingExampleActivity.kt
@@ -27,6 +27,7 @@ class AccountOnboardingExampleActivity : BasicExampleComponentActivity() {
             ViewAccountOnboardingExampleBinding.inflate(LayoutInflater.from(context)).root
                 .apply {
                     initialize(
+                        context = context,
                         embeddedComponentManager = embeddedComponentManager,
                         listener = null,
                         props = props,
