+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/features/payouts/PayoutsExampleActivity.kt
@@ -29,6 +29,7 @@ class PayoutsExampleActivity : BasicExampleComponentActivity() {
             ViewPayoutsExampleBinding.inflate(LayoutInflater.from(context)).root
                 .apply {
                     initialize(
+                        context = context,
                         embeddedComponentManager = embeddedComponentManager,
                         listener = listener,
                         props = EmptyProps,
